fix discussion mailer
