+++ app/views/poll_mailer/poll_option_added_author.html.haml
@@ -1,4 +1,4 @@
 = render "poll_mailer/common/header", poll: @poll, event: @event, action_name: @action_name
 = render "poll_mailer/common/option_added", poll: @poll, event: @event, recipient: @recipient
-= render "poll_mailer/common/vote", poll: @poll, recipient: @recipient, title_key: 'poll_mailer.common.option_added_you_voted'
+-# = render "poll_mailer/common/vote", poll: @poll, recipient: @recipient, title_key: 'poll_mailer.common.option_added_you_voted'
 = render "poll_mailer/common/footer", poll: @poll, recipient: @recipient
